+++ protostar/test_runner/environments/setup_case_execution_environment.py
@@ -2,10 +2,10 @@ from typing import List
 
 from starkware.starknet.business_logic.execution.objects import CallInfo
 
-from protostar.commands.test.fuzzing.strategies import StrategiesHintLocal
 from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet.hint_local import HintLocal
 from protostar.test_runner.cheatcodes import GivenCheatcode, SkipCheatcode
+from protostar.test_runner.fuzzing.strategies import StrategiesHintLocal
 
 from .setup_execution_environment import (
     SetupCheatcodeFactory,
