+++ protostar/test_runner/fuzzing/strategies/integers.py
@@ -3,8 +3,8 @@ from typing import Optional
 from hypothesis.strategies import SearchStrategy, integers
 from starkware.cairo.lang.compiler.ast.cairo_types import CairoType, TypeFelt
 
-from protostar.commands.test.fuzzing.exceptions import SearchStrategyBuildError
-from protostar.commands.test.fuzzing.strategy_descriptor import StrategyDescriptor
+from protostar.test_runner.fuzzing.exceptions import SearchStrategyBuildError
+from protostar.test_runner.fuzzing.strategy_descriptor import StrategyDescriptor
 
 
 class IntegersStrategyDescriptor(StrategyDescriptor):
