+++ protostar/test_runner/fuzzing/strategies/one_of.py
@@ -3,8 +3,8 @@ from typing import Any, cast
 from hypothesis.strategies import SearchStrategy, one_of
 from starkware.cairo.lang.compiler.ast.cairo_types import CairoType
 
-from protostar.commands.test.fuzzing.exceptions import SearchStrategyBuildError
-from protostar.commands.test.fuzzing.strategy_descriptor import StrategyDescriptor
+from protostar.test_runner.fuzzing.exceptions import SearchStrategyBuildError
+from protostar.test_runner.fuzzing.strategy_descriptor import StrategyDescriptor
 
 
 class OneOfStrategyDescriptor(StrategyDescriptor):
