+++ tests/integration/cheatcodes/assume_and_reject/assume_and_reject_test.py
@@ -2,10 +2,9 @@ from pathlib import Path
 
 from starkware.crypto.signature.signature import FIELD_PRIME
 
-from protostar.commands.test.fuzzing.fuzz_input_exception_metadata import (
+from protostar.test_runner.fuzzing.fuzz_input_exception_metadata import (
     FuzzInputExceptionMetadata,
 )
-
 from tests.integration.conftest import (
     RunCairoTestRunnerFixture,
     assert_cairo_test_cases,
